feat(calendar): extract
